epsilon greedy function
