+++ globals.py
@@ -1,5 +1,6 @@
 import os
 
+# GEOSPATIAL DATA
 GEOSPATIAL_BOUND = (59.9112, 59.9438, 10.7027, 10.7772)
 GEOSPATIAL_BOUND_NEW = (59.9040, 59.9547, 10.6478, 10.8095)
 MAX_DISTANCE = 12
@@ -32,7 +33,10 @@ SCOOTER_SPEED = 7
 ITERATION_LENGTH_MINUTES = 20
 NUMBER_OF_ROLLOUTS = 10
 BATTERY_LIMIT = 20.0
+
+# DECISION PARAMETERS
 DISCOUNT_RATE = 0.80
+EPSILON = 0.1  # Probability of taking a random action
 
 # Negative reward for lost trips
 LOST_TRIP_REWARD = -0.1
